Always expose control over `DebugLayers`
